+++ parachain/runtime/snowbase/src/lib.rs
@@ -587,7 +587,7 @@ parameter_types! {
 	pub const MaxVoluntaryExitSize: u32 = 16;
 	pub const MaxAttestationSize: u32 = 128;
 	pub const MaxValidatorsPerCommittee: u32 = 2048;
-	pub const MaxSlotsPerHistoricalRoot: u32 = 64;
+	pub const MaxSlotsPerHistoricalRoot: u64 = 64;
 	pub const MaxFinalizedHeaderSlotArray: u32 = 1000;
 	pub const WeakSubjectivityPeriodSeconds: u32 = 97200;
 	pub const ChainForkVersions: ForkVersions = ForkVersions{
