+++ parachain/runtime/snowblink/src/lib.rs
@@ -585,7 +585,7 @@ parameter_types! {
 	pub const MaxVoluntaryExitSize: u32 = 16;
 	pub const MaxAttestationSize: u32 = 128;
 	pub const MaxValidatorsPerCommittee: u32 = 2048;
-	pub const MaxSlotsPerHistoricalRoot: u32 = 8192;
+	pub const MaxSlotsPerHistoricalRoot: u64 = 8192;
 	pub const MaxFinalizedHeaderSlotArray: u32 = 1000;
 	pub const WeakSubjectivityPeriodSeconds: u32 = 97200;
 	pub const ChainForkVersions: ForkVersions = ForkVersions{
